New LTP documentation
The new LTP documentation is meant to use Sphinx as the main
documentation framework, with ReconStructedText as the main language.
All the conversion has been done step-by-step, each chapter at time,
updating english syntax and fixing typos. There are minor improvements,
but overall the structure is the same.

C / Networking / KVM API are not documented yet, because they will be
probably integrated in the LTP library. C API documentation is already
on going, but the others not.

For a demo, please take a look at:

https://ltp-acerv.readthedocs.io/en/latest/index.html

Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20240320-new_website-v1-0-79b603c8aea1@suse.com



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240314-new_website-3b7505c742e4",
    "prefixes": [],
    "history": {
      "v1": [
        "20240320-new_website-v1-0-79b603c8aea1@suse.com"
      ]
    }
  }
}